requirements.txt
asgiref==3.4.1
astroid==2.8.0
autopep8==1.5.7
cffi==1.15.0
cryptography==35.0.0
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
grpcio==1.41.0
grpcio-tools==1.41.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mock==4.0.3
platformdirs==2.4.0
protobuf==3.18.0
pycodestyle==2.7.0
pycparser==2.21
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uWSGI==2.0.20
wrapt==1.12.1
mysqlclient==2.0.3
